[FeatureExtractor] Create linguistic feature vector
List of features are in the comments of _create_linguistic_features().

Few additions to make this possible. Save the following while creating
contexts.
self.rec_to_utt = {}
self.word_to_syl = {}

Also added 'springy' to cmudict because it's not in there, and same problem as before.

TODO: Create the "position of the current frame of the current phone" feature
for acoustic model, i.e. phoneme is 50ms, linguistic feature is repeated 10 times,
which frame is it currently.