[Networks] Batch data
Pass a closure to the transform parameter in ParallelDatasetIterator in order
to turn the batch input (table of 2D tensors, where 1st dimension is sequence
length (i.e. number of phonemes), 2nd dimension is feature size) into padded
3D tensor. Same thing with the targets.

Padding required because not all sequences are of the same length. We then
use MaskZero in the network to mask.

Test plan: th network.lua
- With lr=0.001, loss goes to down 0.08 or so