[Data][Networks] Re-do acoustic model targets. Train acoustic model
Purpose of this commit:
1) Use the mel-cepstrum coefficients of the spectrograms instead of the
full spectrogram. This changes the dimensionality for both the
spectral envelope and the aperiodicity from 513 to 41.
2) Fix the previous implementation in feature_extractor.py. I forgot to
account for the fact that I ignored the PAU's when extracting phonemes
and calculating utterance duration. To correc this, we look up the start time of
the first non-PAU phoneme and the calculated utterance duration in order to
slice out the appropriate sections of the spectral features.

In order to accomplish 1, I used the Julia wrapper for WORLD. I installed Julia by
installing the dmg and dragging to applications. Packages can be installed using
Pkg inside the Julia interpreter. (Unfortunately the python wrapper for WORLD
doesn't have functions for extracting mel-cepstrum coefficients).

I also checked the reconstructed waveform using the approximated sp and ap, and it is
very very close to identical to the reconstructed waveform using the full sp and ap.
This was done in the example.ipynb in WORLD.jl.

Finally, the logic for stacking input frames for the acoustic model is handled in
dataset.lua. There is now a DurationDataset and a AcousticDataset. Lines 98-117
specifically handle how to account for how many frames each phoneme should be (i.e.
what to do when a phoneme should be a non-integer 12.25 frames).

Test plan:
1) 'th network.lua' with duration model to make sure it still runs and trains
2) 'th network.lua' with acoustic model to check that it runs and trains (starts at
700, decreases)