[Experiment] Train on two speakers
Train on two speakers (slt + clb), both female.

I want to gain some intuition on how necessary it is to have single speaker.
Also, this will be necessary to experiment with the mixture density model.

Results: currently wavs have some snippets that are clearly voices / half-words,
but it seems to be worse than it should be.... At least I think it should be possible
to produce intelligible results. Plus, the loss was slightly lower than when trained on just
one speaker.

Note: super weird BUG
In Network:get_iterator(), sometimes the sample in the transform closure is nil.
It should be a table of size batchsize, but sometimes it returns a combined
tensor instead of a table, i.e. a tensor of size batchsize x .... By sometimes, I mean
it would return a few batches when training the duration model before throwing an error.
Thus, I tried to make sure that the sample was always a table by explicitly writing the
merge closure in tnt.BatchDataset. I swear this solved it and I trained both the duration
and acoustic model like this. However, when I tried testing, I got nil samples again. A
workaround was to bump the batchsize up to 64. Smaller batchsizes did not work....

Test plan:
Local - after changing paths,
1) python feature_extractor.py
2) julia acoustic_feature_extractor.jl
3) scp features to <shannon>:data/processed/cmu_us_clb_arctic...

On shannon:
4) th main.lua -maxepochs 500 -save_model_every_epoch 20 -models_dir models_two_datasets -lr 0.0005 -method adam -model acoustic -notes linear254to512_linear512to512_lstm512to256_lstm256to256_linear256to84__QUINPHONE_f0INTERPLATE -gpuid 2 -two_datasets
5) th main.lua -model duration -notes linear254to256_linear256to128_lstm128to128_linear128to1 -save_model_every_epoch 10 -maxepochs 200 -lr 0.001 -method adam -gpuid 0 -two_datasets -models_dir models_two_datasets
6) th main.lua -gpuid 0 -mode test -load_duration_model_path models_two_datasets/duration/2016_8_9___18_5_43/net_e1000.t7 -load_acoustic_model_path models_two_datasets/acoustic/2016_8_9___17_46_29/net_e160.t7 -save_test_dir outputs_two_datasets/spectral -two_datasets -batchsize 64

Local:
7) scp output_two_datasets/spectral/full_pipeline h5 files over
8) julia generate_test_wavs.jl (after changing paths)