[Blizzard] Parse, create features, and train on Blizzard2013 data
Blizzard2013 is a speech synthesis challenge. It includes ~15-20 hours of
segmented wav files (as well as 300 hours of unsegmented speech).

Notes on parsing prompts:
When parsing and saving prompts, handle OOV words by:
- making first force-align pass to see if any oov words (ProsodyLab-Aligner produces OOV.txt)
- now, using OOV.txt, delete any wavs and labs that contain oov word
- make second-force align pass

**NOTE on silence phones/words:
- ProsodyLab-Aligner alignments include phones and words that are 'sil' in the beginning and end
- However, there are also silences in the middle
- When extracting, for example, the timings for duration targets:
  - the beginning and end silences are ignored
  - the time of the silences in the middle are split between the previous and next phone / word

Notes on refactor to be backwards compatible with CMUArctic
- Paths for models and outputs changed (e.g. instead of models/ and models_twodatasets/ in the
root directory, have models/blizzard/, models/cmuarctic/, and models/cmuarctic_twodatasets/)
- Command line arguments and paths changed in main.lua, network.lua, DataSplitter.lua, dataset.lua

**NOTE: Again, super weird bug in Network:get_iterator()'s transform closure.
The sample returned by dataset.lua is a minibatch, where sample.input and sample.target
are supposed to be tables of tensors, where each tensor is one input and one target. However,
sometimes, it'll train for a while (< 1 epoch) and then throw an error because sample.input and sample.target
are tensors instead of tables, with the first dimension simply being the batchsize (i.e. all the
information is the same, only it's been transformed from a table to a tensor). I faced this error
in an earlier commit and had a work around by increasing the batch size. I did the same thing
this time, but I don't know what the underlying issue is.

Test plan:
- python BlizzardFeatureExtractor.py
- th DataSplitter.lua -dataset blizzard2013
- scp linguistic_inputs, duration_targets, acoustic_targets, data_splits to Shannon
- th main.lua -dataset blizzard2013 -model duration
  - Also acoustic, both locally and on Shannon using GPU

Regression tests on CMUArctic:
- python ArcticFeatureExtractor.py
- julia ArcticAcousticFeatureExtractor.jl
- th DataSplitter.lua -dataset cmuarctic
- th main.lua -dataset cmuarctic -model duration
  - Also acoustic, both locally and on Shannon using GPU